fix(calculate_max_capacity): Fix to prevent mixed shelfs within the same feature-store.
If the mixed depts by feature are allowed. It can cause some combinations to end up with a shelf found for a dept and no shelf found for another one. With this fix, shelfs within the same feature-store are prioritized by the quantity and the result is joined by feature-store.