fix(process-shelf-data): Corrected the file name used to save the shelves data. Deleted unnecessary changes to the data base.
The changes to the "vp" column are unnecessary because that column is not used or displayed in the app. For some reason, these defects don't let the app work correctly in Element production. The app does start but when tried to run the query, it instantly fails, sending the 'DOWNLOAD FAILED' log and the no info downloaded user alert.